refactor: 패키지 구조 개선
- infrastructure 패키지 제거
- global 패키지 추가
- application 패키지 추가
