+++ pekko/src/main/java/com/tok/pekko/application/actor/event/ShutdownEvent.java
@@ -1,0 +1,6 @@
+package com.tok.pekko.application.actor.event;
+
+import com.tok.pekko.global.common.CborSerializable;
+
+public record ShutdownEvent() implements CborSerializable {
+}
