--- pekko/src/main/java/com/tok/pekko/infrastructure/persistence/event/ShutdownEvent.java
@@ -1,6 +1,0 @@
-package com.tok.pekko.infrastructure.persistence.event;
-
-import com.tok.pekko.common.CborSerializable;
-
-public record ShutdownEvent() implements CborSerializable {
-}
