+++ pekko/src/main/resources/application-dev.conf
@@ -4,13 +4,11 @@ pekko {
     provider = cluster
 
     serializers {
-      jackson-json = "org.apache.pekko.serialization.jackson.JacksonJsonSerializer"
       jackson-cbor = "org.apache.pekko.serialization.jackson.JacksonCborSerializer"
     }
 
     serialization-bindings {
-      "com.tok.pekko.common.CborSerializable" = jackson-cbor
-      "java.io.Serializable" = java
+      "com.tok.pekko.global.common.CborSerializable" = jackson-cbor
     }
   }
