+++ pekko/src/test/java/com/tok/pekko/global/config/dev/BlockHoundConfigurationTest.java
@@ -1,4 +1,4 @@
-package com.tok.pekko.infrastructure.config.dev;
+package com.tok.pekko.global.config.dev;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
