+++ pekko/src/test/java/com/tok/pekko/global/config/dev/BlockHoundTestInstallUtils.java
@@ -1,4 +1,4 @@
-package com.tok.pekko.infrastructure.config.dev;
+package com.tok.pekko.global.config.dev;
 
 import reactor.blockhound.BlockHound;
