+++ pekko/src/test/resources/application.conf
@@ -5,7 +5,7 @@ pekko {
     provider = cluster
 
     serialization-bindings {
-      "com.tok.pekko.common.CborSerializable" = jackson-cbor
+      "com.tok.pekko.global.common.CborSerializable" = jackson-cbor
     }
   }
